# swagger/swagger.json
{"swagger":"2.0","info":{"title":"ISprime","description":"This API exposes an image resource that allows uploading and downloading images","contact":{"name":"Mauro Delazeri","email":"mauro.delazeri@image-uploader.com"},"version":""},"host":"localhost:8080","schemes":["http"],"consumes":["application/json"],"produces":["application/json"],"paths":{"/api/images/":{"post":{"tags":["image"],"summary":"upload image","description":"Upload multiple images in multipart request","operationId":"image#upload","produces":["application/json; type=collection"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ImageMediaCollection"}}},"schemes":["http"]}},"/api/images/{id}":{"get":{"tags":["image"],"summary":"show image","description":"Show an image metadata","operationId":"image#show","parameters":[{"name":"id","in":"path","description":"Image ID","required":true,"type":"integer"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ImageMedia"}},"404":{"description":"Not Found"}},"schemes":["http"]}},"/download/{filename}":{"get":{"summary":"Download images/","operationId":"image#/download/*filename","parameters":[{"name":"filename","in":"path","description":"Relative file path","required":true,"type":"string"}],"responses":{"200":{"description":"File downloaded","schema":{"type":"file"}},"404":{"description":"File not found","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/swagger-ui/{filepath}":{"get":{"summary":"Download swagger-ui/","operationId":"swagger#/swagger-ui/*filepath","parameters":[{"name":"filepath","in":"path","description":"Relative file path","required":true,"type":"string"}],"responses":{"200":{"description":"File downloaded","schema":{"type":"file"}},"404":{"description":"File not found","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/swagger.json":{"get":{"summary":"Download swagger/swagger.json","operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}}},"definitions":{"ImageMedia":{"title":"Mediatype identifier: application/json; view=default","type":"object","properties":{"filename":{"type":"string","description":"Image filename","example":"Illo aut."},"id":{"type":"integer","description":"Image ID","example":5810014414692312219,"format":"int64"},"uploaded_at":{"type":"string","description":"Upload timestamp","example":"2014-03-10T11:53:37Z","format":"date-time"}},"description":"Image metadata (default view)","example":{"filename":"Illo aut.","id":5810014414692312219,"uploaded_at":"2014-03-10T11:53:37Z"},"required":["id","filename","uploaded_at"]},"ImageMediaCollection":{"title":"Mediatype identifier: application/json; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/ImageMedia"},"description":"ImageMediaCollection is the media type for an array of ImageMedia (default view)","example":[{"filename":"Illo aut.","id":5810014414692312219,"uploaded_at":"2014-03-10T11:53:37Z"},{"filename":"Illo aut.","id":5810014414692312219,"uploaded_at":"2014-03-10T11:53:37Z"}]},"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}},"responses":{"NotFound":{"description":"Not Found"}}}